titan-ultima/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "titan-ultima"
dynamic = ["version"]
description = "TITAN – Tool for Interpreting and Transforming Archival Nodes. CLI and library for Ultima file formats (U8: Pagan, U7: Black Gate & Serpent Isle)."
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "theGreyWanderer", email = "196163911+theGreyWanderer-uc@users.noreply.github.com" }]
keywords = [
    "ultima", "ultima8", "ultima7", "u8", "u7", "pagan", "black-gate", "serpent-isle",
    "game-modding", "mod-tools", "asset-extraction", "file-formats",
    "flex", "shapes", "xmidi",
    "cli", "batch-processing",
    "retro-gaming", "graphics-conversion", "sound-export",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment :: Role-Playing",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Sound/Audio :: Conversion",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Topic :: File Formats",
    "Typing :: Typed",
]

dependencies = [
    "numpy>=1.24",
    "pillow>=10.0",
    "typer>=0.15",
    "freetype-py>=2.5",
    "ziafont>=0.11",
    'tomli>=2.0; python_version < "3.11"',
]

[project.urls]
Homepage = "https://github.com/theGreyWanderer-uc/tgwUltima"
Repository = "https://github.com/theGreyWanderer-uc/tgwUltima"
Changelog = "https://github.com/theGreyWanderer-uc/tgwUltima/blob/main/titan-ultima/CHANGELOG.md"
"Bug Tracker" = "https://github.com/theGreyWanderer-uc/tgwUltima/issues"

[tool.hatch.version]
path = "src/titan/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/titan"]
# Exclude luts and bundled fonts from package auto-discovery (non-Python data)
# to prevent hatchling from adding binary files twice. Re-include via force-include.
exclude = ["src/titan/adaptive_resample/luts", "src/titan/fonts/bundled"]

[tool.hatch.build.targets.wheel.force-include]
"src/titan/adaptive_resample/luts" = "titan/adaptive_resample/luts"
"src/titan/fonts/bundled" = "titan/fonts/bundled"

[project.scripts]
titan = "titan.cli:main"